Deprecate EmptyFilter in favor of NullFilter
The EmptyFilter was doing a query `IS NULL` or
`IS NOT NULL` so it should be renamed NullFilter
to prevent misundestanding with the query `IS EMPTY`.
