+++ UPGRADE-3.x.md
@@ -1,6 +1,10 @@
 UPGRADE 3.x
 ===========
 
+### Sonata\DoctrineORMAdminBundle\Filter\EmptyFilter
+
+The EmptyFilter is deprecated, use NullFilter instead.
+
 ### Sonata\DoctrineORMAdminBundle\Filter\Filter
 
 Deprecate passing an instance of `Sonata\AdminBundle\Datagrid\ProxyQueryInterface`
